Bug fix for constant realisation cases
Which appears when some individus have the same realisation.
If the out-bootstrapped individus have the same value for one variable
and if they are the only ones out-strapped, then no conclusion can be
taken for the rf$importance[, 1].
The "randomForest" gives Nan values, which were not compatible with the
VSURF algorithm.